Register the "MIDI Device Info..." menu item
